+++ src/sparkledriver/core.clj
@@ -70,7 +70,7 @@
 (defn fetch!
   "Fetch 'url' using 'browser', return browser after loading is complete."
   [browser url]
-  (.get browser url)
+  (.get browser (str url))
   browser)
 
 (defn status-code
@@ -150,7 +150,7 @@
   (.getAttribute element "id"))
 
 (defn text
-  "Return the complete textual content of `element` (including any children)."
+  "Return the complete visible textual content of `element` (including any children). Text from hidden elements is not included."
   [element]
   (.getText element))
 
@@ -349,3 +349,13 @@
             :value  (.getValue %2)})
    {}
    (.getCookies (.manage browser))))
+
+(defn page-text
+  "Return the complete visible textual content of the current page. Text from hidden elements is not included."
+  [browser]
+  (text (find-by-tag browser "html")))
+
+(defn title
+  "Return the title of the current page."
+  [browser]
+  (inner-html (find-by-css browser "head title")))