+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- `page-text` returns the textual content of the current page
+- `title` returns the page's title
+
 ### Changed
+- `fetch!` now calls `str` on its argument, making it compatible with URL types like `java.net.URL` and `lambdaisland.uri.URI`.
 
 ## [0.1.9] - 2017-08-09
 ### Added